requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.11
python-dateutil>=2.8.2
schedule>=1.2.0
plyer>=2.1.0
python-dotenv>=1.0.0
urllib3>=2.0.7 